Fixed homepage to conditionally render into the admin page and added edit book pop up to page
