--- src/components/styles/Admin.css
@@ -1,121 +1,0 @@
-
-
-#cont-ad {
-    max-width: 90%;
-    padding-bottom: 5%;
-}
-
-#title-ad {
-    font-size: large;
-    font-weight: bold;
-    background-color:rgb(61, 65, 65);
-    border: 2px solid white;
-    border-radius: 25px;
-    margin-top: 3%;
-    margin-bottom: 2%;    
-    margin-left: 7%;
-    margin-right: 7%;
-
-}
-
-#col-onsale-ad {
-
-}
-
-#col-bestseller-ad {
-
-}
-
-#promo-bestseller-row-ad {
-    background-image: url('../images/background.jpg');
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-attachment: scroll;
-    border: 2px solid white;
-    border-radius: 25px;
-
-    text-align: center;
-
-    margin: 2%;
-    max-width: 90%;
-    padding-top: 1%;
-    padding-bottom: 2%
-}
-
-#column-ad {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 2%;
-    margin-bottom: 2%;
-    }
-
-#card-style-ad {
-    padding: 3%;
-    width: 100%;
-    height: 100%;
-    max-height: 100%;
-    background-color: #ffffff;
-    }
-
-
-#lGI-title-ad {
-    background-color: transparent;
-    padding-top: 3%;
-}
-
-#lGI-ad {
-    background-color: transparent;
-}
-
-#lGI-links-ad {
-    /* background-color: transparent; */
-    text-align: center;
-}
-
-#but-mb-ad {
-    background-color: transparent;
-    text-align: center;
-    border: none;
-    color: rgb(15, 117, 241);
-
-}
-
-#image-ad {
-    max-height: 45%;
-    max-width: 55%;
-    align-content: center
-}
-
-#adddel-outtercont-ad {
-    display: flex;
-    justify-content: center;
-    padding: 2vh;
-    margin-top: 2vh;
-
-
-}
-
-#adddel-innercont-ad {
-    background-color: rgb(53, 53, 53);
-    max-width: 80vh;
-    min-width: 80vh;
-
-    padding: 2vh;
-    border-radius: 5px;
-    border: 0.5px solid #ffffff;
-}
-
-#buttoncont-ad {
-    display: flex;
-    justify-content: space-around;
-    padding-top: 1.5vh;
-
-
-}
-
-#titlecont-ad {
-    text-align: center;
-    padding-bottom: .5vh;
-    border-bottom: 2px solid black;
-    color: white;
-}
